Implement image data extraction and decompression
This commit implements a method of processing the raw image data
successfully, thus finishing the chunk management part of the program.
It also adds to the project the "kunzip" INFLATE decompression library,
specifically the minimum required files to allow decompression of the
temporary file containing the image data.
Defiltering and interlaced images support yet to be implemented.